Move components implementation to a dedicate folder
