+++ src/tribler-core/tribler_core/components/implementation/bandwidth_accounting.py
