+++ src/tribler-core/tribler_core/components/implementation/popularity.py
