+++ src/tribler-core/tribler_core/components/implementation/resource_monitor.py
