+++ src/tribler-core/tribler_core/components/implementation/watch_folder.py
